Bug/en 2812 round index int64 (#242)
* Replace in32 with int64 when referring to the round index in folder 'data'

* Replace in32 with int64 when referring to the round index in folder 'process'

* Replace in32 with int64 when referring to the round index in folder 'consensus'

* Replace in32 with int64 when referring to the round index in folder 'integrationTests'

* Fix minor typo in variable name

* Updated capnp schema, rounds are now uint64

* Replace more in32 with int64 when referring to the round index in folder 'consensus'

* Replace more in32 with int64 when referring to the round index in folder 'process'

* Replace more in32 with int64 when referring to the round index in folder 'data'

* Replace more int32 with int64 in folder 'integrationTests'

* Replace int32 with int64 when referring to the round in folders 'cmd' and 'node'

* Many updates from int32 round index to int64

* Replace uint32 with uint64 for round in headerHandlerStub.go

* Find more 'round'-related variables of types int32 and uint32 and change them to 64-bit

* Apply feedback from pull request

* Fix minor imports formatting

* Fix integration tests after merging development

+++ consensus/mock/mockTestInitializer.go
@@ -16,7 +16,7 @@ func InitChronologyHandlerMock() consensus.ChronologyHandler {
 
 func InitBlockProcessorMock() *BlockProcessorMock {
 	blockProcessorMock := &BlockProcessorMock{}
-	blockProcessorMock.CreateBlockCalled = func(round uint32, haveTime func() bool) (data.BodyHandler, error) {
+	blockProcessorMock.CreateBlockCalled = func(round uint64, haveTime func() bool) (data.BodyHandler, error) {
 		emptyBlock := make(block.Body, 0)
 
 		return emptyBlock, nil
@@ -28,7 +28,7 @@ func InitBlockProcessorMock() *BlockProcessorMock {
 	blockProcessorMock.ProcessBlockCalled = func(blockChain data.ChainHandler, header data.HeaderHandler, body data.BodyHandler, haveTime func() time.Duration) error {
 		return nil
 	}
-	blockProcessorMock.CreateBlockHeaderCalled = func(body data.BodyHandler, round uint32, haveTime func() bool) (header data.HeaderHandler, e error) {
+	blockProcessorMock.CreateBlockHeaderCalled = func(body data.BodyHandler, round uint64, haveTime func() bool) (header data.HeaderHandler, e error) {
 		return &block.Header{RootHash: []byte{}}, nil
 	}
 	blockProcessorMock.DecodeBlockBodyCalled = func(dta []byte) data.BodyHandler {
@@ -99,7 +99,7 @@ func InitConsensusCore() *ConsensusCoreMock {
 	}
 	blockProcessorMock := InitBlockProcessorMock()
 	blockTrackerMock := &BlocksTrackerMock{
-		SetBlockBroadcastRoundCalled: func(nonce uint64, round int32) {
+		SetBlockBroadcastRoundCalled: func(nonce uint64, round int64) {
 		},
 	}
 	bootstrapperMock := &BootstrapperMock{}
